TestunitFlowCase部分代码提交

+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/SystemDao.java
@@ -15,7 +15,7 @@ public interface SystemDao {
      * @param systemDO
      * @return
      */
-    public int addSystemDO(SystemDO systemDO);
+    public Long addSystemDO(SystemDO systemDO);
 
     /**
      * @param systemDO
