+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/TestunitFlowCaseDao.java
@@ -15,7 +15,7 @@ public interface TestunitFlowCaseDao {
      * @param testunitFlowCaseDO
      * @return
      */
-    public int addTestunitFlowCaseDO(TestunitFlowCaseDO testunitFlowCaseDO);
+    public Long addTestunitFlowCaseDO(TestunitFlowCaseDO testunitFlowCaseDO);
 
     /**
      * @param testunitFlowCaseDO
