+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/TestunitParamDao.java
@@ -15,7 +15,7 @@ public interface TestunitParamDao {
      * @param testunitParamDO
      * @return
      */
-    public int addTestunitParamDO(TestunitParamDO testunitParamDO);
+    public Long addTestunitParamDO(TestunitParamDO testunitParamDO);
 
     /**
      * @param testunitParamDO
