+++ com.alibaba.testme.dao/src/main/java/com/alibaba/testme/dao/WorkSpaceDao.java
@@ -15,7 +15,7 @@ public interface WorkSpaceDao {
      * @param workSpaceDO
      * @return
      */
-    public int addWorkSpaceDO(WorkSpaceDO workSpaceDO);
+    public Long addWorkSpaceDO(WorkSpaceDO workSpaceDO);
 
     /**
      * @param workSpaceDO
