+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitFlowCaseDetailService.java
@@ -15,7 +15,7 @@ public interface TestunitFlowCaseDetailService {
      * @param testunitFlowCaseDetailDO
      * @return
      */
-    public int addTestunitFlowCaseDetailDO(TestunitFlowCaseDetailDO testunitFlowCaseDetailDO);
+    public Long addTestunitFlowCaseDetailDO(TestunitFlowCaseDetailDO testunitFlowCaseDetailDO);
 
     /**
      * @param DO
