+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/TestunitParamExtService.java
@@ -15,7 +15,7 @@ public interface TestunitParamExtService {
      * @param testunitParamExtDO
      * @return
      */
-    public int addTestunitParamExtDO(TestunitParamExtDO testunitParamExtDO);
+    public Long addTestunitParamExtDO(TestunitParamExtDO testunitParamExtDO);
 
     /**
      * @param DO
